backend/requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
Pillow>=10.0.0,<11.0.0  # For image processing
django-cors-headers>=4.0.0,<5.0.0
django-environ>=0.10.0,<1.0.0
whitenoise>=6.4.0,<7.0.0  # For serving static files

# Database
psycopg2-binary>=2.9.5,<3.0.0
redis>=4.5.0,<5.0.0

# API Documentation
drf-yasg>=1.21.0,<2.0.0

# Authentication
djangorestframework-simplejwt>=5.2.2,<6.0.0

# ASGI Server
gunicorn>=20.1.0,<21.0.0
uvicorn>=0.20.0,<0.21.0

# Development
django-debug-toolbar>=3.8.0,<4.0.0
pylint-django>=2.5.0,<3.0.0

# Testing
pytest>=7.3.1,<8.0.0
pytest-django>=4.5.2,<5.0.0
pytest-cov>=4.0.0,<5.0.0

# Code Quality
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.2.0,<2.0.0

django-extensions>=3.2.0,<4.0.0
django-redis>=5.3.0,<6.0.0
